MainWindow: Set up the AvahiBrowser
This is probably a bit weird.  I made it go in this place, because I
think the discovery part should run as soon as possible, i.e. at the
start of the application.  This is debatable, of course.  I'm also
lacking a vision of how to get the data in to the signing page.